setup fitting for looping param in tweezer_statistics.py and moved lorentzian fitting utilities to base_statistics.py
