+++ common/bulk_gas_statistics.py
@@ -51,8 +51,7 @@ class BulkGasStatistician(BaseStatistician):
         self.plot_config = plot_config or PlotConfig()
         self._load_processed_quantities(preproc_h5_path)
         self._save_mloop_params(shot_h5_path)
-        h5p = Path(preproc_h5_path)
-        self.folder_path = h5p.parent
+        self.folder_path = Path(preproc_h5_path).parent
 
     def _load_processed_quantities(self, preproc_h5_path):
         """Load processed quantities from an h5 file.
@@ -332,66 +331,6 @@ class BulkGasStatistician(BaseStatistician):
             self.save_subfig(fig, figname)
         else:
             fig.savefig(figname)
-
-    @staticmethod
-    def save_subfig(subfig, filename):
-        # Create a new figure with the same size as the subfigure
-        new_fig = plt.figure(figsize=subfig.get_figure().get_size_inches())
-
-        # Get the position and size of the subfigure
-        bbox = subfig.get_tightbbox(subfig.figure.canvas.get_renderer())
-        bbox = bbox.transformed(subfig.figure.transFigure.inverted())
-
-        # Create a new axes that fills the entire figure
-        new_axes = new_fig.add_axes([0, 0, 1, 1])
-
-        # Draw the subfigure content
-        new_axes.set_facecolor(subfig.get_facecolor())
-        subfig.figure.canvas.draw()
-
-        # Save the new figure
-        new_fig.savefig(filename, bbox_inches='tight', pad_inches=0)
-        plt.close(new_fig)
-
-    @staticmethod
-    def lorentzian(x, x0, width, a, offset):
-        """
-        Returns a Lorentzian function.
-
-        Parameters
-        ----------
-        x : float or array
-            Input values
-        x0 : float
-            Central frequency
-        width : float
-            Width of the Lorentzian
-        a : float
-            Amplitude of the Lorentzian
-        offset : float
-            Offset of the Lorentzian
-
-        Returns
-        -------
-        float or array
-            Lorentzian function values
-        """
-        detuning = x - x0
-        return a * width / ((width / 2)**2 + detuning**2) + offset
-
-    def fit_lorentzian(self, x_data, y_data):
-        """
-        Fits a Lorentzian function to the atom number data.
-        """
-        x0_guess = x_data[np.argmax(y_data)]
-        x_resolution = (x_data[1] - x_data[0])
-        width_guess = 2*x_resolution
-        y_data_range = np.max(y_data) - np.min(y_data)
-        a_guess = y_data_range / width_guess * (width_guess/2)**2
-
-        offset_guess = np.min(y_data)
-        p0 = [x0_guess, width_guess, a_guess, offset_guess]
-        return optimize.curve_fit(self.lorentzian, x_data, y_data, p0=p0)
 
     @staticmethod
     def time_of_flight_expansion(time, initial_width, temperature):
